+++ R/PSTR.R
@@ -2,8 +2,8 @@
 ## package name: PSTR
 ## author: Yukai Yang
 ## Statistiska Inst., Uppsala Universitet
-## Sep 2017
-################################################################################# 
+## Oct 2017
+#################################################################################
 
 
 #' PSTR: A package implementing the Panel Smooth Transition Regression (PSTR) modelling.
@@ -25,6 +25,11 @@
 #' \deqn{y_{it} = \mu_i + \beta_0' x_{it} + \beta_1' z_{it} g_{it} + u_{it}}
 #' where \eqn{g_{it}} is the transition function taking the logistic form with the transition variable for individual \eqn{i}, \eqn{x_{it}} contains the explanatory variables in the linear part, and \eqn{z_{it}} contains the explanatory variables in the nonlinear part, and they can be different.
 #'
+#' The transition function \eqn{g_{it}} takes the logistic form
+#' \deqn{g(q_{it} ; \gamma, c) = \left( 1 + \exp \left( - \gamma \prod_{j=1}^{m} (q_{it} - c_j) \right) \right)^{-1}}
+#' with \eqn{\gamma > 0} and \eqn{c_1 < c_2 < ... < c_m}. \eqn{\gamma} can be reparametrized as \eqn{\gamma = \exp{\delta}} where \eqn{\delta} is a real number.
+#'
+#'
 #' @section Author and Maintainer:
 #' Yukai Yang
 #'
@@ -33,7 +38,7 @@
 #' \email{yukai.yang@@statistik.uu.se}
 #'
 #' @section References:
-#' González, A., Teräsvirta, T., van Dijk, D. and Yang, Y. (2017) \emph{Panel Smooth Transition Regression Models}
+#' González, A., Teräsvirta, T., van Dijk, D. and Yang, Y. (2005) "\href{http://swopec.hhs.se/hastef/papers/hastef0604.pdf}{Panel Smooth Transition Regression Models}", SSE/EFI Working Paper Series in Economics and Finance 604, Stockholm School of Economics, revised 11 Oct 2017.
 #'
 #' @section Function for Initialization:
 #' \code{\link{NewPSTR}} initialize the modelling by creating an object of the class PSTR.
@@ -57,12 +62,16 @@
 #' \code{\link{version}} shows the version number and some information of the package.
 #'
 #' \code{\link{print.PSTR}} prints the object of the class PSTR.
-#' 
+#'
 #' \code{\link{plot_transition}} plots the transition function of an estimated PSTR model.
-#' 
+#'
+#' \code{\link{plot_response}} plots curve or surfaces of the expected reponse agaist the corresponding variable.
+#'
 #' @section  Data:
 #' \code{\link{Hansen99}} a balanced panel of 565 US firms observed for the years 1973–1987.
-#' 
+#'
+#' \code{\link{sunspot}} transformed Wolf annual sunspot numbers for the years 1710-1979.
+#'
 #' @docType package
 #' @name PSTR
 NULL
@@ -71,10 +80,16 @@ NULL
 #' @importFrom stats optim pchisq pf quantile
 NULL
 
+#' @importFrom magrittr %>%
+NULL
+
 #' @import tibble
 NULL
 
-#' @importFrom ggplot2 ggplot aes geom_point labs scale_x_log10
+#' @importFrom ggplot2 ggplot aes geom_point geom_line labs scale_x_log10
+NULL
+
+#' @importFrom plotly add_surface add_trace plot_ly layout
 NULL
 
 #' @importFrom snowfall sfInit sfExport sfSapply sfStop
@@ -164,4 +179,4 @@ NewPSTR <- function(data, dep, indep, indep_k=NULL, tvars, im=1, iT)
   ret$vYb = vYb; ret$mXb = mXb
 
   return(ret)
-}         
+}