+++ NEWS.md
@@ -1,6 +1,19 @@
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-New Features in PSTR 1.1.0
-==========================
+New Features in PSTR 1.2.1 'Orange Panel'
+=========================================
+
+-   Link to the working paper.
+
+New Features in PSTR 1.2.0 'Orange Panel'
+=========================================
+
+-   New function "plot\_response" which use the ggplot2 and plotly packages to plot the curve or surface of the expected response (contribution) against some variable and the transition variable.
+-   The estimation function "EstPSTR" will return the estimated time-invariante individule effects.
+-   The width of the separating lines in printing the results can adjust itself automatically to fit the width of the console.
+-   Rewrite the "version" function.
+
+New Features in PSTR 1.1.0 'Red Panel'
+======================================
 
 -   Better documentation.
 -   In the function "EstPSTR", now the initial value for "gamma" is allowed. The user can freely choose either "gamma" or "delta" to input for the estimation.
@@ -8,6 +21,7 @@ New Features in PSTR 1.1.0
 -   The initial values can now be chosen automatically and will be returned after estimation.
 -   New function "plot\_transition" which use the ggplot2 package to plot the estimated transition function.
 -   The package ggplot2 imported.
+-   A new data set "sunspot" included.
 -   Some bugs fixed.
 
 New Featurs in PSTR 1.0.1